refactor: extract hydration from hooks
