+++ packages/integration-karma/test-hydration/mismatches/element-instead-of-textNode/index.spec.js
@@ -21,7 +21,7 @@ export default {
 
         expect(consoleCalls.error).toHaveSize(2);
         expect(consoleCalls.error[0][0].message).toContain(
-            '[LWC error]: Hydration mismatch: Unexpected VNode type.'
+            '[LWC error]: Hydration mismatch: incorrect node type received'
         );
         expect(consoleCalls.error[1][0]).toContain('Recovering from error while hydrating');
     },
